refactor coordinateService to coordinate-utils
