+++ projects/laji/src/app/+project-form/results/common/service/result.service.ts
@@ -2,7 +2,7 @@ import { map, tap } from 'rxjs/operators';
 import { Injectable } from '@angular/core';
 import { Observable, Observer, of as ObservableOf } from 'rxjs';
 import { WarehouseApi } from '../../../../shared/api/WarehouseApi';
-import { convertYkjToGeoJsonFeature, CoordinateService } from '../../../../shared/service/coordinate.service';
+import { convertYkjToGeoJsonFeature } from '../../../../root/coordinate-utils';
 import { TaxonomyApi } from '../../../../shared/api/TaxonomyApi';
 import { WarehouseQueryInterface } from '../../../../shared/model/WarehouseQueryInterface';
 
@@ -38,8 +38,7 @@ export class ResultService {
 
   constructor(
     private warehouseApi: WarehouseApi,
-    private taxonomyApi: TaxonomyApi,
-    private coordinateService: CoordinateService
+    private taxonomyApi: TaxonomyApi
   ) { }
 
   getTaxon(taxonId: string) {
